# src/main/resources/static/swagger-ui/openapi.yaml
openapi: 3.0.1
info:
  title: API del Sistema Bancario
  description: API REST para gestión de cuentas bancarias y usuarios.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Servidor local de desarrollo

paths:
  /api/cuentas/crear/{usuarioId}:
    post:
      summary: Crear una cuenta bancaria para un usuario
      parameters:
        - name: usuarioId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CuentaBancariaInput'
      responses:
        '200':
          description: Cuenta creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CuentaBancaria'

  /api/cuentas/usuario/{usuarioId}:
    get:
      summary: Obtener cuentas de un usuario específico
      parameters:
        - name: usuarioId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de cuentas del usuario
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CuentaBancaria'

  /api/usuarios:
    get:
      summary: Listar todos los usuarios
      responses:
        '200':
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Usuario'
    post:
      summary: Crear un nuevo usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioInput'
      responses:
        '200':
          description: Usuario creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'

  /api/transacciones/transferir:
    post:
      summary: Realizar una transferencia entre cuentas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferenciaDTO'
      responses:
        '200':
          description: Transferencia realizada con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaccion'

components:
  schemas:
    CuentaBancaria:
      type: object
      properties:
        id:
          type: integer
        numeroCuenta:
          type: string
        saldo:
          type: number
          format: double
        usuario:
          $ref: '#/components/schemas/Usuario'

    CuentaBancariaInput:
      type: object
      properties:
        # No incluyo id ni usuario porque se gestionan en el backend
        numeroCuenta:
          type: string
        saldo:
          type: number
          format: double

    Usuario:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
        email:
          type: string

    UsuarioInput:
      type: object
      required:
        - nombre
        - email
      properties:
        nombre:
          type: string
        email:
          type: string

    TransferenciaDTO:
      type: object
      required:
        - cuentaOrigenId
        - cuentaDestinoId
        - monto
      properties:
        cuentaOrigenId:
          type: integer
        cuentaDestinoId:
          type: integer
        monto:
          type: number
          format: double

    Transaccion:
      type: object
      properties:
        id:
          type: integer
        fecha:
          type: string
          format: date-time
        monto:
          type: number
          format: double
        cuentaOrigen:
          $ref: '#/components/schemas/CuentaBancaria'
        cuentaDestino:
          $ref: '#/components/schemas/CuentaBancaria'